Adding refseq_crawler.pl
refseq_crawler.pl is a command-line tool for crawling the NCBI website to look for updates to the reference genome data stored in the refseq dataset.  It takes two parameters, the root directory in which the local copy of conp-dataset is stored, and the name of the user whose github fork of conp-dataset it will update and then generate a PR from. Note that running this script requires enough space on the local machine to download any updated genomics files (so that git annex can generate a checksum for these files.)